refactor: move check for duplicate arguments to a separate module
